+++ tests/pipelines/controlnet/test_flax_controlnet.py
@@ -30,7 +30,7 @@ if is_flax_available():
 
 @slow
 @require_flax
-class FlaxStableDiffusionControlNetPipelineIntegrationTests(unittest.TestCase):
+class FlaxControlNetPipelineIntegrationTests(unittest.TestCase):
     def tearDown(self):
         # clean up the VRAM after each test
         super().tearDown()
